requirements.txt
setuptools~=65.5.1
minimalmodbus~=2.1.1
pyserial~=3.5